+++ src/mopidy/audio/__init__.py
@@ -1,8 +1,7 @@
 from mopidy.audio._actor import Audio, AudioProxy
 from mopidy.audio._listener import AudioListener
+from mopidy.audio._utils import supported_uri_schemes
 from mopidy.types import PlaybackState
-
-from .utils import supported_uri_schemes
 
 __all__ = [
     "Audio",