+++ docs/changelog.rst
@@ -131,6 +131,7 @@ Changes to the Audio API may affect a few Mopidy backend extensions.
 
   - :mod:`mopidy.audio.actor`
   - :mod:`mopidy.audio.listener`
+  - :mod:`mopidy.audio.utils`
 
 - Removed APIs only used by Mopidy-Spotify's bespoke audio delivery mechanism,
   which has not been used since Spotify shut down their libspotify APIs in